Extracted CaseTileHelper
